refactor newreservepage folder - move inside components folder
